requirements.txt
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
plotly>=5.0.0
jupyter>=1.0.0
pyserial>=3.5
